Various desgin tweaks (see commit message)
1.Re routed some traces to eliminate some jumpers.
2. Ran ratsnest to fill in ground plane
Note: I ran a DRC and it gave me a bunch of stop errors. This is becouse compontent silkscreens were confilcting with their pads. As we're not using a silk screen this should be fine
Note: In the upper left hand side there's a jumper that looks like it goes no where. This is connedcting the ground plane of the audio jack to the rest of the board as it got enclosed.